Compiler-warning sweep, round 2: clear all remaining 408 lint warnings
Drives `./gradlew compileJava` with `-Xlint:all -Xmaxwarns 9999` from
408 warnings to 0 across the main source set.

- 96 dangling-doc-comments: 81 leading /** copyright headers
  downgraded to /* (still valid block comments, no longer parse as
  dangling Javadoc), and the 15 ///////////// banners in
  NodeTypeBase.java replaced with //----------- (Java 25's JEP 467
  treats /// as Markdown Javadoc).
- 23 lossy-conversions: explicit casts at the assignment site for the
  GUI-int / vertex-float / MjBattery.extractPower-int boundaries.
- 1 fallthrough: @SuppressWarnings on FunctionSignature.parse0.
- Singletons: redundant casts removed from PipeFlowItems (typed
  capability) and MessageVolumeBoxes (readNbt returns CompoundTag);
  static-via-instance fixed in GuiArchitectTable; Guava Charsets.UTF_8
  → StandardCharsets.UTF_8; try-with-resources unnamed variable in
  ScriptableRegistry; @SafeVarargs + @SuppressWarnings("varargs") on
  CombinedItemHandlerWrapper.
- Stale "TEMPORARY DO NOT USE" @Deprecated removed from
  AssemblyRecipe/AssemblyRecipeBasic/StackDefinition (load-bearing,
  not actually temporary), the 3 anonymous FunctionContext
  constructors, and IFakePlayerProvider.getBuildCraftPlayer.
- AssemblyRecipe.setRegistryName(String|Object) → final, so the
  subclass constructor's call no longer escapes this.
- Bulk class-level @SuppressWarnings for the residual vanilla
  deprecations (TextureAtlas, Item, BlockStateModel, BlockStateBase,
  Ingredient, BlockBehaviour, …) and the unavoidable unchecked /
  rawtypes (generic registry/handler patterns, generic array creation).
  Class-level so future categories merge in one place.

Tests pass. One [deprecation] note remains in the test source set
(AdvancementTester) — tracked in todos.md.

+++ src/main/java/buildcraft/factory/tile/TileMiningWell.java
@@ -47,7 +47,7 @@ public class TileMiningWell extends TileMiner {
         if (currentPos != null && canBreak()) {
             shouldCheck = true;
             long target = BlockUtil.computeBlockBreakPower(level, currentPos);
-            progress += battery.extractPower(0, target - progress);
+            progress += (int) battery.extractPower(0, target - progress);
             if (progress >= target) {
                 progress = 0;
                 level.destroyBlockProgress(currentPos.hashCode(), currentPos, -1);
